etl/requirements.txt
ijson==3.3.0
psycopg==3.2.4
typing_extensions==4.12.2
python-dotenv
psycopg2-binary